WIP refactoring detector
* Removed ConstructBeamlineSpecificFGSParams
* Moved gridscan-related setup code into <beamline>/gridscan.py
* Adding detector-specific modules to BeamlineSpecificFGSFeatures
