+++ src/mx_bluesky/hyperion/experiment_plans/hyperion_beamline_specific.py
@@ -2,21 +2,15 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from functools import partial
-from pathlib import Path
 from typing import Any
 
-import bluesky.plan_stubs as bps
 from bluesky.protocols import Readable
 from bluesky.utils import MsgGenerator
 from dodal.devices.eiger import EigerDetector
 
 from mx_bluesky.common.device_setup_plans.eiger import tidy_eiger
 from mx_bluesky.common.device_setup_plans.gridscan import (
-    panda_fast_gridscan_params,
-    set_panda_fgs_params,
     set_zebra_fgs_3d_params,
-)
-from mx_bluesky.common.device_setup_plans.setup_zebra_and_shutter import (
     setup_zebra_for_gridscan,
     tidy_up_zebra_after_gridscan,
 )
@@ -28,17 +22,13 @@ from mx_bluesky.common.experiment_plans.common_flyscan_xray_centre_plan import (
 from mx_bluesky.common.parameters.components import DiffractionExperiment
 from mx_bluesky.common.parameters.device_composites import TDetector
 from mx_bluesky.common.parameters.gridscan import GridScanParams
-from mx_bluesky.common.utils.log import LOGGER
 from mx_bluesky.hyperion.blueapi.composites import (
     HyperionInternalGridDetectThenXRayCentreComposite,
 )
-from mx_bluesky.hyperion.device_setup_plans.setup_panda import (
-    disarm_panda_for_gridscan,
-    set_panda_directory,
-    setup_panda_for_flyscan,
-)
-from mx_bluesky.hyperion.device_setup_plans.setup_zebra import (
-    setup_zebra_for_panda_flyscan,
+from mx_bluesky.hyperion.device_setup_plans.gridscan import (
+    panda_tidy,
+    panda_triggering_setup,
+    set_panda_fgs_params,
 )
 from mx_bluesky.hyperion.external_interaction.config_server import (
     get_hyperion_feature_settings,
@@ -52,7 +42,6 @@ class SmargonSpeedError(Exception):
 def construct_hyperion_specific_features(
     xrc_composite: HyperionInternalGridDetectThenXRayCentreComposite[TDetector],
     xrc_parameters: TSetupParameters,
-    grid_scan_params: GridScanParams,
 ) -> BeamlineSpecificFGSFeatures[
     HyperionInternalGridDetectThenXRayCentreComposite[TDetector], TSetupParameters
 ]:
@@ -89,19 +78,20 @@ def construct_hyperion_specific_features(
     ]
 
     if get_hyperion_feature_settings().USE_PANDA_FOR_GRIDSCAN:
-        setup_trigger_plan = _panda_triggering_setup
-        tidy_plan = partial(_panda_tidy, xrc_composite)
+        setup_trigger_plan = partial(
+            panda_triggering_setup, settings=get_hyperion_feature_settings()
+        )
+        tidy_plan = partial(panda_tidy, xrc_composite)
         set_flyscan_params_plan = partial(
             set_panda_fgs_params,
             xrc_composite.panda_fast_grid_scan,
             xrc_parameters,
-            set_stub_offsets=get_hyperion_feature_settings().SET_STUB_OFFSETS,
-            run_up_distance_mm=get_hyperion_feature_settings().PANDA_RUNUP_DISTANCE_MM,
+            settings=get_hyperion_feature_settings(),
         )
         fgs_motors = xrc_composite.panda_fast_grid_scan
 
     else:
-        setup_trigger_plan = _setup_zebra_for_gridscan
+        setup_trigger_plan = setup_zebra_for_gridscan
         tidy_plan = partial(
             tidy_up_zebra_after_gridscan,
             xrc_composite.zebra,
@@ -129,12 +119,6 @@ def construct_hyperion_specific_features(
     return features
 
 
-def _setup_zebra_for_gridscan(
-    composite: HyperionInternalGridDetectThenXRayCentreComposite, _, __
-):
-    yield from setup_zebra_for_gridscan(composite)
-
-
 def _detector_signals_to_read_during_collection(detector: Any) -> list[Readable]:
     match detector:
         case EigerDetector():
@@ -145,72 +129,3 @@ def _detector_signals_to_read_during_collection(detector: Any) -> list[Readable]
             ]
         case _:
             raise ValueError(f"Unsupported detector type {type(detector)}")
-
-
-def _panda_tidy(xrc_composite: HyperionInternalGridDetectThenXRayCentreComposite):
-    group = "panda_flyscan_tidy"
-    LOGGER.info("Disabling panda blocks")
-    yield from disarm_panda_for_gridscan(xrc_composite.panda, group)
-    yield from tidy_up_zebra_after_gridscan(
-        xrc_composite.zebra, xrc_composite.sample_shutter, group=group, wait=False
-    )
-    yield from bps.unstage(xrc_composite.panda, group=group)
-    yield from bps.wait(group, timeout=10)
-
-
-def _panda_triggering_setup(
-    xrc_composite: HyperionInternalGridDetectThenXRayCentreComposite,
-    parameters: DiffractionExperiment,
-    grid_scan_parameters: GridScanParams,
-) -> MsgGenerator:
-    LOGGER.info("Setting up Panda for flyscan")
-
-    run_up_distance_mm = yield from bps.rd(
-        xrc_composite.panda_fast_grid_scan.run_up_distance_mm
-    )
-
-    detector_deadtime_s = 1e-4  # This value was empirically found to be safer than the documented deadtime in the Eiger manual
-
-    time_between_x_steps_ms = (detector_deadtime_s + parameters.exposure_time_s) * 1e3
-
-    smargon_speed_limit_mm_per_s = yield from bps.rd(xrc_composite.gonio.x.max_velocity)
-
-    panda_params = panda_fast_gridscan_params(parameters, grid_scan_parameters)
-    sample_velocity_mm_per_s = (
-        panda_params.x_step_size_mm * 1e3 / time_between_x_steps_ms
-    )
-    if sample_velocity_mm_per_s > smargon_speed_limit_mm_per_s:
-        raise SmargonSpeedError(
-            f"Smargon speed was calculated from x step size\
-            {panda_params.x_step_size_mm}mm and\
-            time_between_x_steps_ms {time_between_x_steps_ms} as\
-            {sample_velocity_mm_per_s}mm/s. The smargon's speed limit is\
-            {smargon_speed_limit_mm_per_s}mm/s."
-        )
-    else:
-        LOGGER.info(
-            f"Panda grid scan: Smargon speed set to {sample_velocity_mm_per_s} mm/s"
-            f" and using a run-up distance of {run_up_distance_mm}"
-        )
-
-    yield from bps.mv(
-        xrc_composite.panda_fast_grid_scan.time_between_x_steps_ms,
-        time_between_x_steps_ms,
-    )
-
-    directory_provider_root = Path(parameters.storage_directory)
-    yield from set_panda_directory(directory_provider_root)
-
-    yield from setup_panda_for_flyscan(
-        xrc_composite.panda,
-        panda_params,
-        xrc_composite.gonio,
-        parameters.exposure_time_s,
-        time_between_x_steps_ms,
-        sample_velocity_mm_per_s,
-    )
-
-    LOGGER.info("Setting up Zebra for panda flyscan")
-    yield from setup_zebra_for_panda_flyscan(
-        xrc_composite.zebra, xrc_composite.sample_shutter, wait=True
-    )
